move peer manager class to lbrynet.core
