+++ lbrynet/tests/functional/test_reflector.py
@@ -4,8 +4,7 @@ from twisted.trial import unittest
 from lbrynet import conf
 from lbrynet.core.StreamDescriptor import get_sd_info
 from lbrynet import reflector
-from lbrynet.core import BlobManager
-from lbrynet.dht import peermanager
+from lbrynet.core import BlobManager, PeerManager
 from lbrynet.core import Session
 from lbrynet.core import StreamDescriptor
 from lbrynet.lbry_file.client import EncryptedFileOptions
@@ -28,7 +27,7 @@ class TestReflector(unittest.TestCase):
         self.port = None
         self.addCleanup(self.take_down_env)
         wallet = mocks.Wallet()
-        peer_manager = peermanager.PeerManager()
+        peer_manager = PeerManager.PeerManager()
         peer_finder = mocks.PeerFinder(5553, peer_manager, 2)
         hash_announcer = mocks.Announcer()
         sd_identifier = StreamDescriptor.StreamDescriptorIdentifier()
